refactor: centralize interactive runtime operations
